Background image didn't show on Chrome 53
In old Chrome 53, it didn't know how to render a background image url + a css-variable-based HSLA color value in a radial-gradient on one DOM node and would just show 2 of the 3 backgrounds being applied. This moves the url to a separate variable and applies that to a new background layer (pseudo-class).
Bonus, now the URL background can be used in other places too.

Signed-off-by: Blake Stephens <blake.stephens@lge.com>