Added additional supported screen resolutions
These allow for easier testing to keep the display scaled proportionally when testing or doing screen recording. The Quarter FHD ("qtrfhd") lets you circumvent a device emulation bug in Chrome when testing or working with the drag-an-drop API not respecting the device scaling during a drag. This also lets a Retina™ display capture video at "FHD" resolution, due to the doubled device-pixel-ratio.

Signed-off-by: Blake Stephens <blake.stephens@lge.com>