Button size are quite small on skins Carbon, Cobalt, Copper, Electro, Titanium. Since there is no Ui guide, it would be good to look the same as Gallium.
Solution:
@agate-mediacontrols-icon-font-size:  inherit => @agate-icon-font-size;

Shuffle button active is too close to the color of the symbol on skins:
Cobalt and Copper(only on night mode)

Solution:
Added skin prop in MediaPlayer.js that contains the name of
the current skin, it is passed on in MediaControls.js.There
it checks of the current skin is Cobalt or Copper.If no the
css class of Button Shuffle is css.repeat, else it is
css.exception.

Buttons are too close to one another and they overlap with the progress
bar on skins: Electro and Titanium.
The play button overlaps with the progress bar bubble on Carbon skin.

Solution:
@agate-mediacontrols-button-height: inherit => none;
@agate-mediacontrols-button-margin: inherit => 0 30px 0 30px;
@agate-mediacontrols-button-min-width: inherit => none;
@agate-mediaplayer-times-padding: inherit => @agate-component-spacing;
@agate-mediacontrols-icon-font-size: inherit => @agate-icon-font-size;
@agate-mediacontrols-playpause-icon-font-size: inherit => @agate-icon-font-size;
@agate-mediacontrols-playpause-button-height: inherit => @agate-icon-font-size * 1.1;

Solution:
Changed the MediaSlider.module.less and added the
background-color property
to both .slider.fill and .focus.fill.
Added a new variable in colors-base.less called
@agate-mediaslider-fill-bg-color with the value: inherit;

Added to
// MediaSlider
@agate-mediaslider-fill-bg-color: @agate-accent;
@agate-mediaslider-focused-bar-color: @agate-highlight;